chore: updates with the latest EDC refactor
