+++ core/core-utils/src/main/java/org/eclipse/tractusx/edc/core/utils/credentials/CredentialTypePredicate.java
@@ -19,7 +19,7 @@
 
 package org.eclipse.tractusx.edc.core.utils.credentials;
 
-import org.eclipse.edc.identitytrust.model.VerifiableCredential;
+import org.eclipse.edc.iam.identitytrust.spi.model.VerifiableCredential;
 
 import java.util.function.Predicate;
