+++ edc-extensions/cx-policy/src/main/java/org/eclipse/tractusx/edc/policy/cx/dismantler/DismantlerCredentialConstraintFunction.java
@@ -20,7 +20,7 @@
 package org.eclipse.tractusx.edc.policy.cx.dismantler;
 
 import jakarta.json.JsonObject;
-import org.eclipse.edc.identitytrust.model.VerifiableCredential;
+import org.eclipse.edc.iam.identitytrust.spi.model.VerifiableCredential;
 import org.eclipse.edc.policy.engine.spi.PolicyContext;
 import org.eclipse.edc.policy.model.Operator;
 import org.eclipse.edc.policy.model.Permission;
