+++ edc-extensions/cx-policy/src/test/java/org/eclipse/tractusx/edc/policy/cx/CredentialFunctions.java
@@ -19,9 +19,10 @@
 
 package org.eclipse.tractusx.edc.policy.cx;
 
-import org.eclipse.edc.identitytrust.model.CredentialSubject;
-import org.eclipse.edc.identitytrust.model.Issuer;
-import org.eclipse.edc.identitytrust.model.VerifiableCredential;
+
+import org.eclipse.edc.iam.identitytrust.spi.model.CredentialSubject;
+import org.eclipse.edc.iam.identitytrust.spi.model.Issuer;
+import org.eclipse.edc.iam.identitytrust.spi.model.VerifiableCredential;
 
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
