+++ edc-extensions/iatp/tx-iatp/src/test/java/org/eclipse/tractusx/edc/iam/iatp/identity/IatpIdentityExtractorTest.java
@@ -19,9 +19,9 @@
 
 package org.eclipse.tractusx.edc.iam.iatp.identity;
 
-import org.eclipse.edc.identitytrust.model.CredentialSubject;
-import org.eclipse.edc.identitytrust.model.Issuer;
-import org.eclipse.edc.identitytrust.model.VerifiableCredential;
+import org.eclipse.edc.iam.identitytrust.spi.model.CredentialSubject;
+import org.eclipse.edc.iam.identitytrust.spi.model.Issuer;
+import org.eclipse.edc.iam.identitytrust.spi.model.VerifiableCredential;
 import org.eclipse.edc.spi.EdcException;
 import org.eclipse.edc.spi.iam.ClaimToken;
 import org.junit.jupiter.api.Test;
